CapstoneProject-Spoken_claim_verification/whisper_service/requirements.txt
openai-whisper==20231117
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.3.0
python-dotenv==1.0.0
